Moved StepikDispatcher to a new module.
